[fixed] Issue in TiledLoader where TMX files stored in CSV format and included rotated or flipped tiles would fail to load.

+++ gdx/src/com/badlogic/gdx/graphics/g2d/tiled/TiledLoader.java
@@ -464,7 +464,7 @@ public class TiledLoader {
 					StringTokenizer st = new StringTokenizer(dataString.trim(), ",");
 					for (int row = 0; row < currLayerHeight; row++) {
 						for (int col = 0; col < currLayerWidth; col++) {
-							currLayer.tiles[row][col] = Integer.parseInt(st.nextToken().trim());
+							currLayer.tiles[row][col] = (int) Long.parseLong(st.nextToken().trim());
 						}
 					}
 				}
